refactor/#45: 패키지 구조 변경 admin -> admin/user
