+++ src/test/java/com/dodo/dodoserver/domain/admin/user/service/AdminServiceTest.java
@@ -1,10 +1,11 @@
-package com.dodo.dodoserver.domain.admin.service;
+package com.dodo.dodoserver.domain.admin.user.service;
 
-import com.dodo.dodoserver.domain.admin.dao.SanctionHistoryRepository;
-import com.dodo.dodoserver.domain.admin.dao.UserAdminRepository;
-import com.dodo.dodoserver.domain.admin.dto.UserSanctionRequestDto;
-import com.dodo.dodoserver.domain.admin.entity.SanctionHistory;
-import com.dodo.dodoserver.domain.admin.entity.SanctionType;
+import com.dodo.dodoserver.domain.admin.user.dao.SanctionHistoryRepository;
+import com.dodo.dodoserver.domain.admin.user.dao.UserAdminRepository;
+import com.dodo.dodoserver.domain.admin.user.dto.UserSanctionRequestDto;
+import com.dodo.dodoserver.domain.admin.user.entity.SanctionHistory;
+import com.dodo.dodoserver.domain.admin.user.entity.SanctionType;
+import com.dodo.dodoserver.domain.admin.user.service.AdminService;
 import com.dodo.dodoserver.domain.user.entity.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
